+++ CHANGELOG.md
@@ -8,12 +8,15 @@ AdminNeo 5.1.0
 
 - Add defaultServer and defaultDatabase configuration parameters
 - Allow creating generated columns (by @vrana)
+- MySQL: Display converting function for binary, bit or geometry fields (by @vrana)
 - MS SQL: Add support for computed columns (by @vrana)
 - SQLite: Add support for generated columns (by @vrana)
 - PostgreSQL: Add support for indexes on materialized views (by @vrana)
+- Add support for CockroachDB (by @vrana)
 - Refine and standardize Japanese translation (by @yama)
 - Plugins: Use protected visibility of member properties
 - Force inline editing with the Shift key when clicking on a link.
+- Editor: Display database views
 
 ### Bugfixes